requirements.txt
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pymupdf
beautifulsoup4
requests